pyproject.toml
[project]
name = "youtube-content-pipeline"
version = "0.4.0"
description = "YouTube content extraction pipeline with transcription and MongoDB storage"
requires-python = ">=3.12"
dependencies = [
    "requests>=2.31.0",
    "youtube-transcript-api>=0.6.2",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "motor>=3.3.0",
    "pymongo>=4.6.0",
    "typer>=0.9.0",
    "rich>=13.7.0",
    "aiohttp>=3.9.0",
    "yt-dlp>=2023.12.30",
    "whisper>=1.1.10",
    "numpy>=2.0.0,<2.1.0",
    "torch>=2.10.0",
    "transformers>=4.36.0,<4.40.0",
    "optimum==1.24.0",
    "optimum-intel[openvino]>=1.20.0,<1.27.0",
    "openvino-genai>=2025.0.0",
    "openvino>=2025.0.0",
    "openvino-tokenizers>=2025.0.0",
    "librosa>=0.10.0",
    "onnx>=1.15.0",
    "onnxruntime>=1.15.0",
    "setuptools<70",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "python-multipart>=0.0.20",
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pyyaml>=6.0",
    "browser-cookie3>=0.19.0",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.1.0",
    "mypy>=1.7.0",
    "pre-commit>=3.6.0",
]
mcp = [
    "mcp>=1.0.0",
    "fastmcp>=0.1.0",
]
redis = [
    "redis>=5.0.0",
    "hiredis>=2.3.0",
]
prometheus = [
    "prometheus-fastapi-instrumentator>=6.1.0",
    "prometheus-client>=0.19.0",
]
rate-limit = [
    "slowapi>=0.1.9",
]

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B", "C4", "SIM"]
ignore = [
    "B008",  # Function call in argument defaults (required for FastAPI/Typer DI)
]

[tool.mypy]
python_version = "3.12"
strict = true
warn_return_any = true
warn_unused_ignores = true

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
